Reorganized to put startup items under imports/startup.
Modeled on the structure of the example Meteor todos app.
