+++ imports/startup/client/index.js
@@ -1,0 +1,1 @@
+import './routes.js';
