+++ imports/startup/server/index.js
@@ -1,0 +1,2 @@
+// Initialize some default data on startup
+import './fixtures.js';
